Create README_WCRs_to_GMS.md
This code can be used to convert Well Completion Reports from California Open Data and organize and feed them into GMS-Borehole Module.
The Steps and marked down in the Jupyter Notebook code. 
Horizons are arbitrary and not accurate, you need to modify them later.
SOIL-IDs need to be decided in advance. I used a look-up table in Excel to assign SOIL-IDs according to soil descriptions.